+++ boot/signalsigterm.go
@@ -4,7 +4,7 @@
 
 // +build darwin dragonfly freebsd linux netbsd openbsd solaris
 
-package main
+package boot
 
 import (
 	"os"
